requirements.txt
python-dotenv~=0.20.0
rich~=12.4.4